2227 - Reorganize WorkflowNodesTabs
