+++ client/src/pages/platform/workflow-editor/components/WorkflowNodesPopoverMenuComponentList.tsx
@@ -1,5 +1,5 @@
 import {Input} from '@/components/ui/input';
-import WorkflowNodesTabs from '@/pages/platform/workflow-editor/components/WorkflowNodesTabs';
+import WorkflowNodesTabs from '@/pages/platform/workflow-editor/components/workflow-nodes-tabs/WorkflowNodesTabs';
 import useWorkflowDataStore from '@/pages/platform/workflow-editor/stores/useWorkflowDataStore';
 import CopilotButton from '@/shared/components/copilot/CopilotButton';
 import {Source} from '@/shared/components/copilot/stores/useCopilotStore';
