+++ client/src/pages/platform/workflow-editor/components/workflow-nodes-tabs/WorkflowNodesTabs.tsx
@@ -1,13 +1,12 @@
-import {ScrollArea} from '@/components/ui/scroll-area';
-import {Tabs, TabsContent, TabsList, TabsTrigger} from '@/components/ui/tabs';
+import {Tabs, TabsList, TabsTrigger} from '@/components/ui/tabs';
 import {ComponentDefinitionBasic, TaskDispatcherDefinition} from '@/shared/middleware/platform/configuration';
 import {useFeatureFlagsStore} from '@/shared/stores/useFeatureFlagsStore';
 import {ClickedDefinitionType} from '@/shared/types';
 import {useEffect, useMemo, useState} from 'react';
 
-import {useComponentFiltering} from '../hooks/useComponentFiltering';
-import ActionComponentsFilter from './ActionComponentsFilter';
-import WorkflowNodesTabsItem from './WorkflowNodesTabsItem';
+import {useComponentFiltering} from '../../hooks/useComponentFiltering';
+import ActionComponentsFilter from '../ActionComponentsFilter';
+import WorkflowNodesTabContent from './WorkflowNodesTabContent';
 
 type DefinitionType = (ComponentDefinitionBasic | TaskDispatcherDefinition) & {
     taskDispatcher: boolean;
@@ -27,46 +26,6 @@ interface WorkflowNodesTabsProps {
     taskDispatcherDefinitions: Array<TaskDispatcherDefinition>;
     triggerComponentDefinitions: Array<ComponentDefinitionBasic>;
 }
-
-interface WorkflowNodesTabContentProps {
-    emptyMessage: string;
-    items: Array<ComponentDefinitionBasic | DefinitionType> | undefined;
-    itemsDraggable: boolean;
-    onItemClick?: (clickedItem: ClickedDefinitionType) => void;
-    selectedComponentName?: string;
-    tabValue: string;
-}
-
-const WorkflowNodesTabContent = ({
-    emptyMessage,
-    items,
-    itemsDraggable,
-    onItemClick,
-    selectedComponentName,
-    tabValue,
-}: WorkflowNodesTabContentProps) => {
-    return (
-        <ScrollArea className="overflow-y-auto px-3">
-            <TabsContent className="mt-0 w-full flex-1" value={tabValue}>
-                <ul className="space-y-2" role="list">
-                    {(!items || items.length === 0) && (
-                        <span className="block px-3 py-2 text-xs text-content-neutral-secondary">{emptyMessage}</span>
-                    )}
-
-                    {items?.map((item) => (
-                        <WorkflowNodesTabsItem
-                            draggable={itemsDraggable}
-                            handleClick={() => onItemClick && onItemClick(item as ClickedDefinitionType)}
-                            key={item.name}
-                            node={item as DefinitionType}
-                            selected={tabValue === 'components' ? selectedComponentName === item.name : undefined}
-                        />
-                    ))}
-                </ul>
-            </TabsContent>
-        </ScrollArea>
-    );
-};
 
 const WorkflowNodesTabs = ({
     actionComponentDefinitions,
